Move all Pages underneath /pages/ route.
